Improve request speed; gracefully handle API limit reached
Previously I had a fudge factor of N-1 for whatever the rate
limit remaining was, to avoid unintentionally hitting the limit.
This removes that limit, but makes each API request explicitly
handle a 429 and retry after the reset time has been reached.

I can appreciate the irony of me being a good netizen against
an actively hostile API service (which inexplicably still
exists as of this date).